requirements.txt
Flask
Flask_SQLAlchemy
mysql-connector-python
gunicorn
python-dotenv